Updated `4spacesdk/ci4authextension` from v1.2.1 to v1.2.2
Fix Volume validation for "nfs_server" and "nfs_path" [Issue #61](https://github.com/4spacesdk/kubernetes-service-orchestrator/issues/61)

+++ ci4/app/Config/Database.php
@@ -85,7 +85,7 @@ class Database extends Config
 
         $this->default['DSN'] = getenv('DB_DSN') ? getenv('DB_DSN') : '';
         $this->default['hostname'] = getenv('DB_HOST');
-        $this->default['port'] = getenv('DB_PORT') ? getenv('DB_PORT') : 3306;
+        $this->default['port'] = getenv('DB_PORT') ? (int)getenv('DB_PORT') : 3306;
         $this->default['database'] = getenv('DB_NAME');
         $this->default['username'] = getenv('DB_USER');
         $this->default['password'] = getenv('DB_PASS');
